Add support for when inline data is not supported
This commit adds support for when using inline data
is not an option.  I also moved the max_inline field
from the uct_rc_verbs_iface_common struct to the
uct_rc_iface struct because I was not able to cleanly
access it from the methods where the method flags were
being set.

It was suggested I use the ucs_derived_of method to get
access to the uct_rc_verbs_iface struct, however, it
appears it is possible I could get either the dc struct,
or the rc struct.  This uncertainty is why I moved the
max_inline field, if this is problematic, please let me
know.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>